CHANGE - Made a temporary fix for STOMP Messages > 8192, by adding a fallback controller; Preparing for adding spring security to enhance webSocket users.

+++ src/main/java/service/impl/AfterTestsServiceImpl.java
@@ -145,7 +145,9 @@ public class AfterTestsServiceImpl implements AfterTestsService {
     }
 
     public void deleteTestOutputResults() throws IOException {
-        LOGGER.info("Deleting recent test source directory: " + testSourcePath.toAbsolutePath().toString());
-        FilesAndDirectoryUtil.deleteDirectory(testSourcePath);
+        if (Files.exists(testSourcePath)) {
+            LOGGER.info("Deleting recent test source directory: " + testSourcePath.toAbsolutePath().toString());
+            FilesAndDirectoryUtil.deleteDirectory(testSourcePath);
+        }
     }
 }
